Rename _nvdi to _index_channels
